working on statistics

+++ src/main/java/org/com/service/TicketService.java
@@ -1,31 +1,37 @@
 package org.com.service;
-
-
 import org.com.entity.*;
 import org.com.exceptions.TicketException;
 import org.com.repository.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 import java.util.Optional;
 
 @Service
 public class TicketService {
 
-    private final TicketRepository ticketRepository;
+    @Autowired
+    private TicketRepository ticketRepository;
 
     @Autowired
-    public TicketService(TicketRepository ticketRepository) {
-        this.ticketRepository = ticketRepository;
-    }
+    private ValidationService validationService;
+
 
     public List<Ticket> findTicketsByUsername(String username) {
         return ticketRepository.findTicketByUsername(username);
     }
 
     public void deleteTicketById(long ticketId) {
-        ticketRepository.deleteById(ticketId);
+        Optional<Ticket> ticket = ticketRepository.findById(ticketId);
+        if (ticket.isPresent()) {
+            Cinema cinema = ticket.get().getCinema();
+            boolean isCinemaExpired = validationService.isCinemaExpired(cinema);
+            if (!isCinemaExpired)
+                ticketRepository.deleteById(ticketId);
+            else
+                throw new TicketException("This ticket can't be refunded, because the date is expired");
+        }
+
     }
 
     public Optional<Ticket> findById(long id) {
@@ -43,8 +49,7 @@ public class TicketService {
         } catch (Exception e) {
             throw new TicketException("one or more entities were not found");
         }
-
-
     }
+
 
 }
